Set up cracker class
